inference/requirements.txt
redis==5.0.0
mlflow==3.6.0
pandas>=2.1.4
scikit-learn>=1.3.0
joblib==1.3.2
pyarrow==14.0.2
kafka-python==2.0.2
boto3
numpy<2
